pyproject.toml
[project]
name = "Stack_Scout"
version = "0.1.3"
description = "An automated job scraper that finds the top 5 latest remote dev jobs matching your stack (React, Python, Node.js, Docker, etc.) from sites like Google Jobs, Remote OK & more"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "beautifulsoup4>=4.13.4",
    "bs4>=0.0.2",
    "fake-useragent>=2.2.0",
    "fastapi>=0.116.1",
    "httpx>=0.28.1",
    "jinja2>=3.1.6",
    "pandas>=2.3.1",
    "pytest>=8.4.1",
    "python-dotenv>=1.1.1",
    "requests>=2.32.4",
    "selenium>=4.34.2",
    "selenium-stealth>=1.0.6",
    "uvicorn>=0.35.0",
    "webdriver-manager>=4.0.2",
    "playwright>=1.35.0",
    "lxml>=6.0.0",
    "pytest-asyncio>=1.1.0",
    "python-multipart>=0.0.20",
    "sqlalchemy>=2.0.42",
    "psycopg2-binary>=2.9.10",
    "psycopg2>=2.9.10",
    "google-generativeai>=0.8.5",
    "email-validator>=2.2.0",
    "bcrypt>=4.3.0",
]
